requirements.txt
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
auto-labeling-pipeline==0.1.23
billiard==4.1.0
blinker==1.6.3
blis==0.7.11
boto3==1.28.73
botocore==1.31.73
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.3
cryptography==41.0.5
cymem==2.0.8
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.8
Django==4.2.6
django-allauth==0.52.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-crispy-forms==2.1
django-drf-filepond==0.5.0
django-filter==22.1
django-health-check==3.17.0
django-plotly-dash==2.2.2
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
doccano==1.8.4
dpd-components==0.1.0
drf-yasg==1.21.7
environs==9.5.0
et-xmlfile==1.1.0
filetype==1.2.0
Flask==3.0.0
flower==1.2.0
furl==2.1.3
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
gunicorn==20.1.0
humanize==4.8.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kombu==5.3.2
langcodes==3.3.0
lml==0.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
murmurhash==1.0.10
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging==23.2
pandas==1.5.3
plotly==5.18.0
polars==0.19.11
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
seqeval==1.2.2
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.49
sqlparse==0.4.4
srsly==2.4.8
tenacity==8.2.3
texttable==1.7.0
thinc==8.2.1
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
waitress==2.1.2
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.3
Werkzeug==3.0.1
whitenoise==6.6.0
